chore(registration-react): Move all registration-api modules under registration-api directory under api
