+++ applications/registration-react/src/api/registration-api/catalogs.js
