+++ applications/registration-react/src/pages/api-import-page/api-import-page.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import _ from 'lodash';
 import { readAsText } from 'promise-file-reader';
 
-import { postApi } from '../../api/api-registration-api';
+import { postApi } from '../../api/registration-api/apis';
 import localization from '../../lib/localization';
 import { ImportDialog } from './import-dialog/import-dialog.component';
 import { ImportFileUpload } from './import-file-upload/import-file-upload.component';
