+++ applications/registration-react/src/pages/api-registration-page/import-dialog/import-dialog.component.jsx
@@ -6,7 +6,7 @@ import _ from 'lodash';
 import { readAsText } from 'promise-file-reader';
 
 import localization from '../../../lib/localization';
-import { patchApi } from '../../../api/api-registration-api';
+import { patchApi } from '../../../api/registration-api/apis';
 import { InputFile } from '../../../components/input-file/input-file.component';
 import { ImportLinkUpload } from '../import-link-upload/import-link-upload.component';
