+++ applications/registration-react/src/pages/dataset-list-page/form-catalog/async-validate-catalog.js
@@ -1,5 +1,5 @@
 import { CATALOG_SUCCESS } from '../../../redux/modules/catalog';
-import { saveCatalog } from '../../../api/dataset-registration-api';
+import { saveCatalog } from '../../../api/registration-api/datasets';
 
 const catalogSaveThunk = ({ catalog }) => async dispatch => {
   const savedCatalog = await saveCatalog(catalog);
