+++ applications/registration-react/src/redux/modules/apiCatalogs.js
@@ -1,6 +1,6 @@
 import _ from 'lodash';
 import { fetchApisIfNeededAction } from './apis';
-import { getApiCatalog, postApiCatalog } from '../../api/api-registration-api';
+import { getApiCatalog, postApiCatalog } from '../../api/registration-api/apis';
 
 export const API_CATALOG_REQUEST = 'API_CATALOG_REQUEST';
 export const API_CATALOG_SUCCESS = 'API_CATALOG_SUCCESS';
